+++ src/util.h
@@ -7,7 +7,7 @@
 
 /**
  * Server/client environment: argument handling, config file parsing,
- * logging, thread wrappers
+ * thread wrappers
  */
 #ifndef BITCOIN_UTIL_H
 #define BITCOIN_UTIL_H
@@ -16,6 +16,7 @@
 #include "config/pivx-config.h"
 #endif
 
+#include "logging.h"
 #include "compat.h"
 #include "tinyformat.h"
 #include "utiltime.h"
@@ -46,91 +47,15 @@ extern int keysLoaded;
 extern bool fSucessfullyLoaded;
 extern std::vector<int64_t> obfuScationDenominations;
 extern std::string strBudgetMode;
-extern std::atomic<uint32_t> logCategories;
+
 extern std::map<std::string, std::string> mapArgs;
 extern std::map<std::string, std::vector<std::string> > mapMultiArgs;
-extern bool fPrintToConsole;
-extern bool fPrintToDebugLog;
+
 extern std::string strMiscWarning;
-extern bool fLogTimestamps;
-extern bool fLogIPs;
-extern volatile bool fReopenDebugLog;
+
 
 void SetupEnvironment();
 bool SetupNetworking();
-
-struct CLogCategoryActive
-{
-    std::string category;
-    bool active;
-};
-
-namespace BCLog {
-    enum LogFlags : uint32_t {
-        NONE        = 0,
-        NET         = (1 <<  0),
-        TOR         = (1 <<  1),
-        MEMPOOL     = (1 <<  2),
-        HTTP        = (1 <<  3),
-        BENCH       = (1 <<  4),
-        ZMQ         = (1 <<  5),
-        DB          = (1 <<  6),
-        RPC         = (1 <<  7),
-        ESTIMATEFEE = (1 <<  8),
-        ADDRMAN     = (1 <<  9),
-        SELECTCOINS = (1 << 10),
-        REINDEX     = (1 << 11),
-        CMPCTBLOCK  = (1 << 12),
-        RAND        = (1 << 13),
-        PRUNE       = (1 << 14),
-        PROXY       = (1 << 15),
-        MEMPOOLREJ  = (1 << 16),
-        LIBEVENT    = (1 << 17),
-        COINDB      = (1 << 18),
-        QT          = (1 << 19),
-        LEVELDB     = (1 << 20),
-        STAKING     = (1 << 21),
-        MASTERNODE  = (1 << 22),
-        MNBUDGET    = (1 << 23),
-        LEGACYZC    = (1 << 24),
-        ALL         = ~(uint32_t)0,
-    };
-}
-
-/** Return true if log accepts specified category */
-static inline bool LogAcceptCategory(uint32_t category)
-{
-    return (logCategories.load(std::memory_order_relaxed) & category) != 0;
-}
-
-/** Returns a string with the log categories. */
-std::string ListLogCategories();
-/** Returns a vector of the active log categories. */
-std::vector<CLogCategoryActive> ListActiveLogCategories();
-/** Return true if str parses as a log category and set the flags in f */
-bool GetLogCategory(uint32_t *f, const std::string *str);
-/** Send a string to the log output */
-int LogPrintStr(const std::string& str);
-
-/** Get format string from VA_ARGS for error reporting */
-template<typename... Args> std::string FormatStringFromLogArgs(const char *fmt, const Args&... args) { return fmt; }
-
-#define LogPrintf(...) do {                                                         \
-    std::string _log_msg_; /* Unlikely name to avoid shadowing variables */         \
-    try {                                                                           \
-        _log_msg_ = tfm::format(__VA_ARGS__);                                       \
-    } catch (tinyformat::format_error &e) {                                               \
-        /* Original format string will have newline so don't add one here */        \
-        _log_msg_ = "Error \"" + std::string(e.what()) + "\" while formatting log message: " + FormatStringFromLogArgs(__VA_ARGS__); \
-    }                                                                               \
-    LogPrintStr(_log_msg_);                                                         \
-} while(0)
-
-#define LogPrint(category, ...) do {                                                \
-    if (LogAcceptCategory((category))) {                                            \
-        LogPrintf(__VA_ARGS__);                                                     \
-    }                                                                               \
-} while(0)
 
 template<typename... Args>
 bool error(const char* fmt, const Args&... args)
@@ -163,9 +88,7 @@ void ReadConfigFile(std::map<std::string, std::string>& mapSettingsRet, std::map
 boost::filesystem::path GetSpecialFolderPath(int nFolder, bool fCreate = true);
 #endif
 boost::filesystem::path GetTempPath();
-boost::filesystem::path GetDebugLogPath();
-bool OpenDebugLog();
-void ShrinkDebugFile();
+
 void runCommand(std::string strCommand);
 
 inline bool IsSwitchChar(char c)
